fixed bookin route
