+++ routes/__init__.py
@@ -1,5 +1,6 @@
 from utils.config import IS_DEV
 
+import routes.bookings
 import routes.mediations
 import routes.offers
 import routes.offerers
